+++ src/main/java/net/mindoth/spellmaker/util/spellform/SingleTargetAtRangeForm.java
@@ -12,7 +12,7 @@ import java.util.LinkedHashMap;
 import java.util.List;
 
 public class SingleTargetAtRangeForm extends AbstractSpellForm {
-    public SingleTargetAtRangeForm(int cost) {
+    public SingleTargetAtRangeForm(float cost) {
         super(cost);
     }
